CAMEL-19279: more config elements to config class
